requirements.txt
openai
langchain
python-dotenv
selenium
unstructured
tiktoken
streamlit
transformers
huggingface-hub
sentence-transformers
unstructured
faiss-cpu
pymupdf
pinecone-client